requirements.txt
Django
gunicorn
celery
redis
psycopg2-binary
python-dotenv
pandas
scikit-learn==1.8.0
google-genai
dj-database-url